chore(capacity): organize unit tests into separate files (#986)
# Goal
The goal of this PR is to move Capacity related unit tests into their
respective files

Closes #971

# Checklist
- [x] Capacity type method unit tests to
`pallets/capacity/src/type_tests.rs`
- [x] Extrinsic function unit tests to
`pallets/capacity/src/extrinsic_tests.rs`
- [x] Helper function unit tests to
`pallets/capacity/src/helper_tests.rs`
- [x] Remove `pallets/capacity/src/tests.rs`
