+++ pallets/capacity/src/extrinsics_tests.rs
@@ -1,10 +1,11 @@
 use super::*;
+use crate as pallet_capacity;
 use crate::mock::*;
-use frame_support::{assert_noop, assert_ok};
-use testing_utils::run_to_block;
+use frame_support::{assert_noop, assert_ok, BoundedVec};
+use testing_utils::{register_provider, run_to_block, staking_events};
 
 #[test]
-fn happy_path() {
+fn withdraw_unstaked_happy_path() {
 	new_test_ext().execute_with(|| {
 		// set up staker and staking account
 		let staker = 500;
@@ -35,7 +36,7 @@ fn happy_path() {
 }
 
 #[test]
-fn correctly_sets_new_lock_state() {
+fn withdraw_unstaked_correctly_sets_new_lock_state() {
 	new_test_ext().execute_with(|| {
 		let staker = 500;
 		let mut staking_account = StakingAccountDetails::<Test>::default();
@@ -58,7 +59,7 @@ fn correctly_sets_new_lock_state() {
 }
 
 #[test]
-fn cleans_up_storage_and_removes_all_locks_if_no_stake_left() {
+fn withdraw_unstaked_cleans_up_storage_and_removes_all_locks_if_no_stake_left() {
 	new_test_ext().execute_with(|| {
 		let mut staking_account = StakingAccountDetails::<Test>::default();
 		let staking_amount: BalanceOf<Test> = 10;
@@ -80,7 +81,7 @@ fn cleans_up_storage_and_removes_all_locks_if_no_stake_left() {
 }
 
 #[test]
-fn cannot_withdraw_if_no_unstaking_chunks() {
+fn withdraw_unstaked_cannot_withdraw_if_no_unstaking_chunks() {
 	new_test_ext().execute_with(|| {
 		let staker = 500;
 		let mut staking_account = StakingAccountDetails::<Test>::default();
@@ -93,7 +94,7 @@ fn cannot_withdraw_if_no_unstaking_chunks() {
 	})
 }
 #[test]
-fn cannot_withdraw_if_unstaking_chunks_not_thawed() {
+fn withdraw_unstaked_cannot_withdraw_if_unstaking_chunks_not_thawed() {
 	new_test_ext().execute_with(|| {
 		let staker = 500;
 		let mut staking_account = StakingAccountDetails::<Test>::default();
@@ -114,11 +115,415 @@ fn cannot_withdraw_if_unstaking_chunks_not_thawed() {
 }
 
 #[test]
-fn error_if_not_a_staking_account() {
+fn withdraw_unstaked_error_if_not_a_staking_account() {
 	new_test_ext().execute_with(|| {
 		assert_noop!(
 			Capacity::withdraw_unstaked(RuntimeOrigin::signed(999)),
 			Error::<Test>::NotAStakingAccount
 		);
 	})
+}
+
+#[test]
+fn stake_works() {
+	new_test_ext().execute_with(|| {
+		let account = 200;
+		let target: MessageSourceId = 1;
+		let amount = 5;
+		register_provider(target, String::from("Foo"));
+		assert_ok!(Capacity::stake(RuntimeOrigin::signed(account), target, amount));
+
+		// Check that StakingAccountLedger is updated.
+		assert_eq!(Capacity::get_staking_account_for(account).unwrap().total, amount);
+		assert_eq!(Capacity::get_staking_account_for(account).unwrap().active, amount);
+		assert_eq!(Capacity::get_staking_account_for(account).unwrap().unlocking.len(), 0);
+
+		// Check that StakingTargetLedger is updated.
+		assert_eq!(Capacity::get_target_for(account, target).unwrap().amount, amount);
+		assert_eq!(Capacity::get_target_for(account, target).unwrap().capacity, amount);
+
+		// Check that CapacityLedger is updated.
+		assert_eq!(Capacity::get_capacity_for(target).unwrap().remaining, amount);
+		assert_eq!(Capacity::get_capacity_for(target).unwrap().total_available, amount);
+		assert_eq!(Capacity::get_capacity_for(target).unwrap().last_replenished_epoch, 1);
+
+		let events = staking_events();
+		assert_eq!(events.first().unwrap(), &Event::Staked { account, target, amount });
+
+		assert_eq!(Balances::locks(&account)[0].amount, amount);
+		assert_eq!(Balances::locks(&account)[0].reasons, WithdrawReasons::all().into());
+	});
+}
+
+#[test]
+fn stake_errors_invalid_target_when_target_is_not_registered_provider() {
+	new_test_ext().execute_with(|| {
+		let account = 100;
+		let target: MessageSourceId = 1;
+		let amount = 1;
+		assert_noop!(
+			Capacity::stake(RuntimeOrigin::signed(account), target, amount),
+			Error::<Test>::InvalidTarget
+		);
+	});
+}
+
+#[test]
+fn stake_errors_insufficient_staking_amount_when_staking_below_minimum_staking_amount() {
+	new_test_ext().execute_with(|| {
+		let account = 100;
+		let target: MessageSourceId = 1;
+		let amount = 1;
+		register_provider(target, String::from("Foo"));
+		assert_noop!(
+			Capacity::stake(RuntimeOrigin::signed(account), target, amount),
+			Error::<Test>::InsufficientStakingAmount
+		);
+	});
+}
+
+#[test]
+fn stake_errors_zero_amount_not_allowed() {
+	new_test_ext().execute_with(|| {
+		let account = 100;
+		let target: MessageSourceId = 1;
+		let amount = 0;
+		assert_noop!(
+			Capacity::stake(RuntimeOrigin::signed(account), target, amount),
+			Error::<Test>::ZeroAmountNotAllowed
+		);
+	});
+}
+
+#[test]
+fn stake_increase_stake_amount_works() {
+	new_test_ext().execute_with(|| {
+		let account = 200;
+		let target: MessageSourceId = 1;
+		let initial_amount = 5;
+		register_provider(target, String::from("Foo"));
+
+		// First Stake
+		assert_ok!(Capacity::stake(RuntimeOrigin::signed(account), target, initial_amount));
+		let events = staking_events();
+		assert_eq!(
+			events.first().unwrap(),
+			&Event::Staked { account, target, amount: initial_amount }
+		);
+
+		assert_eq!(Balances::locks(&account)[0].amount, 5);
+		assert_eq!(Balances::locks(&account)[0].reasons, WithdrawReasons::all().into());
+
+		run_to_block(2);
+
+		let additional_amount = 10;
+		// Additional Stake
+		assert_ok!(Capacity::stake(RuntimeOrigin::signed(account), target, additional_amount));
+
+		// Check that StakingAccountLedger is updated.
+		assert_eq!(Capacity::get_staking_account_for(account).unwrap().total, 15);
+		assert_eq!(Capacity::get_staking_account_for(account).unwrap().active, 15);
+		assert_eq!(Capacity::get_staking_account_for(account).unwrap().unlocking.len(), 0);
+
+		// Check that StakingTargetLedger is updated.
+		assert_eq!(Capacity::get_target_for(account, target).unwrap().amount, 15);
+		assert_eq!(Capacity::get_target_for(account, target).unwrap().capacity, 15);
+
+		// Check that CapacityLedger is updated.
+		assert_eq!(Capacity::get_capacity_for(target).unwrap().remaining, 15);
+		assert_eq!(Capacity::get_capacity_for(target).unwrap().total_available, 15);
+		assert_eq!(Capacity::get_capacity_for(target).unwrap().last_replenished_epoch, 2);
+
+		let events = staking_events();
+		assert_eq!(
+			events.first().unwrap(),
+			&Event::Staked { account, target, amount: additional_amount }
+		);
+
+		assert_eq!(Balances::locks(&account)[0].amount, 15);
+		assert_eq!(Balances::locks(&account)[0].reasons, WithdrawReasons::all().into());
+	});
+}
+
+#[test]
+fn stake_multiple_accounts_can_stake_to_the_same_target() {
+	new_test_ext().execute_with(|| {
+		new_test_ext().execute_with(|| {
+			let target: MessageSourceId = 1;
+			register_provider(target, String::from("Foo"));
+			let account_1 = 200;
+			let stake_amount_1 = 5;
+
+			assert_ok!(Capacity::stake(RuntimeOrigin::signed(account_1), target, stake_amount_1));
+
+			// Check that StakingAccountLedger is updated.
+			assert_eq!(Capacity::get_staking_account_for(account_1).unwrap().total, 5);
+			assert_eq!(Capacity::get_staking_account_for(account_1).unwrap().active, 5);
+			assert_eq!(Capacity::get_staking_account_for(account_1).unwrap().unlocking.len(), 0);
+
+			// Check that StakingTargetLedger is updated.
+			assert_eq!(Capacity::get_target_for(account_1, target).unwrap().amount, 5);
+			assert_eq!(Capacity::get_target_for(account_1, target).unwrap().capacity, 5);
+
+			// Check that CapacityLedger is updated.
+			assert_eq!(Capacity::get_capacity_for(target).unwrap().remaining, 5);
+			assert_eq!(Capacity::get_capacity_for(target).unwrap().total_available, 5);
+			assert_eq!(Capacity::get_capacity_for(target).unwrap().last_replenished_epoch, 1);
+
+			run_to_block(2);
+
+			let account_2 = 300;
+			let stake_amount_2 = 10;
+
+			assert_ok!(Capacity::stake(RuntimeOrigin::signed(account_2), target, stake_amount_2));
+
+			// Check that StakingAccountLedger is updated.
+			assert_eq!(Capacity::get_staking_account_for(account_2).unwrap().total, 10);
+			assert_eq!(Capacity::get_staking_account_for(account_2).unwrap().active, 10);
+			assert_eq!(Capacity::get_staking_account_for(account_2).unwrap().unlocking.len(), 0);
+
+			// Check that StakingTargetLedger is updated.
+			assert_eq!(Capacity::get_target_for(account_2, target).unwrap().amount, 10);
+			assert_eq!(Capacity::get_target_for(account_2, target).unwrap().capacity, 10);
+
+			// Check that CapacityLedger is updated.
+			assert_eq!(Capacity::get_capacity_for(target).unwrap().remaining, 15);
+			assert_eq!(Capacity::get_capacity_for(target).unwrap().total_available, 15);
+			assert_eq!(Capacity::get_capacity_for(target).unwrap().last_replenished_epoch, 2);
+		});
+	});
+}
+
+#[test]
+fn stake_an_account_can_stake_to_multiple_targets() {
+	new_test_ext().execute_with(|| {
+		let target_1: MessageSourceId = 1;
+		let target_2: MessageSourceId = 2;
+		register_provider(target_1, String::from("Foo"));
+		register_provider(target_2, String::from("Boo"));
+
+		let account = 200;
+		let amount_1 = 10;
+		let amount_2 = 7;
+
+		assert_ok!(Capacity::stake(RuntimeOrigin::signed(account), target_1, amount_1));
+		assert_eq!(Capacity::get_staking_account_for(account).unwrap().total, 10);
+
+		run_to_block(2);
+		assert_ok!(Capacity::stake(RuntimeOrigin::signed(account), target_2, amount_2));
+
+		// Check that StakingAccountLedger is updated.
+		assert_eq!(Capacity::get_staking_account_for(account).unwrap().total, 17);
+		assert_eq!(Capacity::get_staking_account_for(account).unwrap().active, 17);
+		assert_eq!(Capacity::get_staking_account_for(account).unwrap().unlocking.len(), 0);
+
+		// Check that StakingTargetLedger is updated for target 1.
+		assert_eq!(Capacity::get_target_for(account, target_1).unwrap().amount, 10);
+		assert_eq!(Capacity::get_target_for(account, target_1).unwrap().capacity, 10);
+
+		// Check that StakingTargetLedger is updated for target 2.
+		assert_eq!(Capacity::get_target_for(account, target_2).unwrap().amount, 7);
+		assert_eq!(Capacity::get_target_for(account, target_2).unwrap().capacity, 7);
+
+		// Check that CapacityLedger is updated for target 1.
+		assert_eq!(Capacity::get_capacity_for(target_1).unwrap().remaining, 10);
+		assert_eq!(Capacity::get_capacity_for(target_1).unwrap().total_available, 10);
+		assert_eq!(Capacity::get_capacity_for(target_1).unwrap().last_replenished_epoch, 1);
+
+		// Check that CapacityLedger is updated for target 2.
+
+		assert_eq!(Capacity::get_capacity_for(target_2).unwrap().remaining, 7);
+		assert_eq!(Capacity::get_capacity_for(target_2).unwrap().total_available, 7);
+		assert_eq!(Capacity::get_capacity_for(target_2).unwrap().last_replenished_epoch, 2);
+	});
+}
+
+#[test]
+fn stake_when_staking_amount_is_greater_than_free_balance_it_stakes_maximum() {
+	new_test_ext().execute_with(|| {
+		let target: MessageSourceId = 1;
+		register_provider(target, String::from("Foo"));
+		let account = 100;
+		// An amount greater than the free balance
+		let amount = 13;
+
+		assert_ok!(Capacity::stake(RuntimeOrigin::signed(account), target, amount));
+
+		// Check that StakingAccountLedger is updated.
+		assert_eq!(Capacity::get_staking_account_for(account).unwrap().total, 10);
+		assert_eq!(Capacity::get_staking_account_for(account).unwrap().active, 10);
+		assert_eq!(Capacity::get_staking_account_for(account).unwrap().unlocking.len(), 0);
+
+		// Check that StakingTargetLedger is updated.
+		assert_eq!(Capacity::get_target_for(account, target).unwrap().amount, 10);
+		assert_eq!(Capacity::get_target_for(account, target).unwrap().capacity, 10);
+
+		// Check that CapacityLedger is updated.
+		assert_eq!(Capacity::get_capacity_for(target).unwrap().remaining, 10);
+		assert_eq!(Capacity::get_capacity_for(target).unwrap().total_available, 10);
+		assert_eq!(Capacity::get_capacity_for(target).unwrap().last_replenished_epoch, 1);
+	});
+}
+
+#[test]
+fn unstake_happy_path() {
+	new_test_ext().execute_with(|| {
+		let token_account = 200;
+		let target: MessageSourceId = 1;
+		let staking_amount = 10;
+		let unstaking_amount = 5;
+
+		register_provider(target, String::from("Test Target"));
+
+		assert_ok!(Capacity::stake(RuntimeOrigin::signed(token_account), target, staking_amount));
+		assert_ok!(Capacity::unstake(
+			RuntimeOrigin::signed(token_account),
+			target,
+			unstaking_amount
+		));
+
+		// Assert that staking account detail values are decremented correctly after unstaking
+		let staking_account_details = Capacity::get_staking_account_for(token_account).unwrap();
+
+		assert_eq!(staking_account_details.unlocking.len(), 1);
+		let mut chunks: BoundedVec<
+			UnlockChunk<BalanceOf<Test>, <Test as frame_system::Config>::BlockNumber>,
+			<Test as pallet_capacity::Config>::MaxUnlockingChunks,
+		> = BoundedVec::default();
+
+		chunks
+			.try_push(UnlockChunk::<BalanceOf<Test>, <Test as frame_system::Config>::BlockNumber> {
+				value: BalanceOf::<Test>::from(5u64),
+				thaw_at: <Test as frame_system::Config>::BlockNumber::from(3u64),
+			})
+			.expect("try_push failed");
+
+		assert_eq!(
+			staking_account_details,
+			StakingAccountDetails::<Test> {
+				active: BalanceOf::<Test>::from(5u64),
+				total: BalanceOf::<Test>::from(10u64),
+				unlocking: chunks,
+			}
+		);
+
+		// Assert that staking target detail values are decremented correctly after unstaking
+		let staking_target_details = Capacity::get_target_for(token_account, target).unwrap();
+
+		assert_eq!(
+			staking_target_details,
+			StakingTargetDetails::<BalanceOf<Test>> {
+				amount: BalanceOf::<Test>::from(5u64),
+				capacity: BalanceOf::<Test>::from(5u64),
+			}
+		);
+
+		// Assert that the capacity detail values for the target are decremented properly after unstaking
+		let capacity_details = Capacity::get_capacity_for(target).unwrap();
+
+		assert_eq!(
+			capacity_details,
+			CapacityDetails::<BalanceOf<Test>, <Test as frame_system::Config>::BlockNumber> {
+				remaining: BalanceOf::<Test>::from(10u64),
+				total_tokens_staked: BalanceOf::<Test>::from(5u64),
+				total_available: BalanceOf::<Test>::from(5u64),
+				last_replenished_epoch: <Test as frame_system::Config>::BlockNumber::from(1u64),
+			}
+		);
+	});
+}
+
+#[test]
+fn unstake_errors_unstaking_amount_is_zero() {
+	new_test_ext().execute_with(|| {
+		let token_account = 200;
+		let target: MessageSourceId = 1;
+		let staking_amount = 10;
+		let unstaking_amount = 0;
+
+		register_provider(target, String::from("Test Target"));
+
+		assert_ok!(Capacity::stake(RuntimeOrigin::signed(token_account), target, staking_amount));
+		assert_noop!(
+			Capacity::unstake(RuntimeOrigin::signed(token_account), target, unstaking_amount),
+			Error::<Test>::UnstakedAmountIsZero
+		);
+	});
+}
+
+#[test]
+fn unstake_errors_max_unlocking_chunks_exceeded() {
+	new_test_ext().execute_with(|| {
+		let token_account = 200;
+		let target: MessageSourceId = 1;
+		let staking_amount = 10;
+		let unstaking_amount = 1;
+
+		register_provider(target, String::from("Test Target"));
+
+		assert_ok!(Capacity::stake(RuntimeOrigin::signed(token_account), target, staking_amount));
+
+		for _n in 0..<Test as pallet_capacity::Config>::MaxUnlockingChunks::get() {
+			assert_ok!(Capacity::unstake(
+				RuntimeOrigin::signed(token_account),
+				target,
+				unstaking_amount
+			));
+		}
+
+		assert_noop!(
+			Capacity::unstake(RuntimeOrigin::signed(token_account), target, unstaking_amount),
+			Error::<Test>::MaxUnlockingChunksExceeded
+		);
+	});
+}
+
+#[test]
+fn unstake_errors_amount_to_unstake_exceeds_amount_staked() {
+	new_test_ext().execute_with(|| {
+		let token_account = 200;
+		let target: MessageSourceId = 1;
+		let staking_amount = 10;
+		let unstaking_amount = 11;
+
+		register_provider(target, String::from("Test Target"));
+
+		assert_ok!(Capacity::stake(RuntimeOrigin::signed(token_account), target, staking_amount));
+		assert_noop!(
+			Capacity::unstake(RuntimeOrigin::signed(token_account), target, unstaking_amount),
+			Error::<Test>::AmountToUnstakeExceedsAmountStaked
+		);
+	});
+}
+
+#[test]
+fn unstake_errors_invalid_target() {
+	new_test_ext().execute_with(|| {
+		let token_account = 200;
+		let target: MessageSourceId = 2;
+
+		let unstaking_amount = 11;
+
+		assert_noop!(
+			Capacity::unstake(RuntimeOrigin::signed(token_account), target, unstaking_amount),
+			Error::<Test>::InvalidTarget
+		);
+	});
+}
+
+#[test]
+fn unstake_errors_not_a_staking_account() {
+	new_test_ext().execute_with(|| {
+		let token_account = 200;
+		let target: MessageSourceId = 1;
+
+		let unstaking_amount = 11;
+
+		register_provider(target, String::from("Test Target"));
+
+		assert_noop!(
+			Capacity::unstake(RuntimeOrigin::signed(token_account), target, unstaking_amount),
+			Error::<Test>::StakingAccountNotFound
+		);
+	});
 }
